Initial commit, adds readme.md, ex8.rb, also a test
I set the textwidth to 72 to see if that works and now I'm going to
ramble on for quite a bit to make sure it's working. Now you get to read
some lines from Milton's Paradise Lost: I should be much for open Warr,
O Perrs, as not behind in hate; if what was urg'd main reason to
perswade immddiate Warr, did not disswade me most and seem to cast
ominous conjecture on the whole success: when he who most excels in fact
of Arms, in what he counsels and in what excels mistrusful, grounds his
courage on despair.

It appears to have worked...

[#95933776]